requirements.txt
awscli-local==0.22.2
black==26.1.0
boto3==1.42.38
botocore==1.42.38
click==8.3.1
colorama==0.4.6
flake8==7.3.0
jmespath==1.1.0
localstack-client==2.11
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
